Moved location of outageChecker.js
