+++ backend/outageChecker.js
